requirements.txt
Django==4.0.6
Pillow==9.2.0
PyYAML==6.0
black==22.6.0
celery==5.2.7
dj-database-url==0.5.0
dj-rest-auth==2.2.5
django-allauth==0.51.0
django-anymail==8.6
django-cors-headers==3.13.0
djangorestframework==3.13.1
factory-boy==3.2.1
gunicorn==20.1.0
python-dotenv==0.20.0
uritemplate==4.1.1
whitenoise==6.2.0
